+++ packages/firefuel/lib/src/firefuel_collection.dart
@@ -393,6 +393,51 @@ abstract class FirefuelCollection<T extends Serializable>
   }
 
   @override
+  Future<void> updateFields({
+    required DocumentId docId,
+    required Map<String, Object?> fields,
+  }) async {
+    await untypedRef.doc(docId.docId).update(fields);
+
+    return;
+  }
+
+  @override
+  Future<void> arrayUnion({
+    required DocumentId docId,
+    required String field,
+    required List<Object?> values,
+  }) {
+    return updateFields(
+      docId: docId,
+      fields: {field: FieldValue.arrayUnion(values)},
+    );
+  }
+
+  @override
+  Future<void> arrayRemove({
+    required DocumentId docId,
+    required String field,
+    required List<Object?> values,
+  }) {
+    return updateFields(
+      docId: docId,
+      fields: {field: FieldValue.arrayRemove(values)},
+    );
+  }
+
+  @override
+  Future<void> serverTimestamp({
+    required DocumentId docId,
+    required String field,
+  }) {
+    return updateFields(
+      docId: docId,
+      fields: {field: FieldValue.serverTimestamp()},
+    );
+  }
+
+  @override
   Future<T> updateOrCreate({
     required DocumentId docId,
     required T value,